Addition of Delete, get, and getAll Employee information
Delete: Called Firestore database then called Employee Collection to reference the employee with the same name. Afterwards deleted the employee from the Collectiion / data base.

Get: Called Firestore, referenced the collection Employee, and the employee by name. Converted the Document reference to Document Snapshot and then to an Employee Object, returned this object.

Delete: Accessed Firestore database, Referenced Employee Collection and retrieved list of documents. Then created an iterator that can iterate through the Document Reference List which point to the individual Document Snapshots. While loop permits the List referenced to be converted to Employee Object and saved to a new List and then it is returned.